pyproject.toml
[project]
name = "autopics"
version = "0.1.0"
description = "Generated Image by DALL-E Model"
authors = [{ name = "Yuki.N", email = "Syukkic@Gmail.com" }]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0.post1",
    "openai>=1.3.6",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.scripts]
dev = 'uvicorn api.main:app --reload'
